pyproject.toml
[project]
name = "cyberfusion-sdk-rs03"
version = "0.1.0"
description = "Python SDK for controlling RobStride RS03 actuators over CAN bus"
authors = [
    {name = "Your Name", email = "you@example.com"},
]
requires-python = ">=3.8"
dependencies = [
    "python-dotenv",
    "python-can",
    "robstride @ git+https://github.com/sirwart/robstride",
]

[project.optional-dependencies]
ros2 = [
    "rclpy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["examples", "rs03_driver"]

[dependency-groups]
dev = []
